+++ cmake/ExternLibs.cmake
@@ -41,7 +41,7 @@ function(FETCHSHAREDTYPES)
 	FetchContent_Declare(
 	  shared_types
 	  GIT_REPOSITORY https://github.com/Anticarium/Shared_Types.git
-	  GIT_TAG        v1.2.0
+	  GIT_TAG        v1.3.0
 	)
 
 	FetchContent_GetProperties(shared_types)