+++ src/anticarium_desktop/MainWindowManager.cpp
@@ -7,7 +7,6 @@ MainWindowManager::MainWindowManager(QObject* parent) : QObject(parent) {
 }
 
 void MainWindowManager::sendData(const shared_types::Control& control) {
-    this->control = control;
     emit sendDataEvent(control);
 }
 
@@ -15,11 +14,11 @@ void MainWindowManager::initialize() {
     JTTP* jttp         = JTTP::instance();
     QTimer* fetchTimer = new QTimer(this);
     connect(this, qOverload<const shared_types::Control&>(&MainWindowManager::sendDataEvent), jttp, qOverload<const shared_types::Control&>(&JTTP::onSendData));
-    connect(this, qOverload<const shared_types::RegimeName&>(&MainWindowManager::sendDataEvent), jttp, qOverload<const shared_types::RegimeName&>(&JTTP::onSendData));
+    connect(this, qOverload<const shared_types::RegimeId&>(&MainWindowManager::sendDataEvent), jttp, qOverload<const shared_types::RegimeId&>(&JTTP::onSendData));
     connect(fetchTimer, &QTimer::timeout, jttp, [&]() { emit requestDataEvent(JTTP::REQUEST_DATA::SENSOR_DATA); });
     connect(this, &MainWindowManager::requestDataEvent, jttp, &JTTP::onRequestData);
     connect(jttp, qOverload<const shared_types::Regimes&>(&JTTP::dataReceivedEvent), this, qOverload<const shared_types::Regimes&>(&MainWindowManager::displayDataEvent));
-    connect(jttp, qOverload<const shared_types::RegimeName&>(&JTTP::dataReceivedEvent), this, qOverload<const shared_types::RegimeName&>(&MainWindowManager::displayDataEvent));
+    connect(jttp, qOverload<const shared_types::RegimeId&>(&JTTP::dataReceivedEvent), this, qOverload<const shared_types::RegimeId&>(&MainWindowManager::displayDataEvent));
     connect(jttp, qOverload<const shared_types::SensorData&>(&JTTP::dataReceivedEvent), this, qOverload<const shared_types::SensorData&>(&MainWindowManager::displayDataEvent));
     connect(jttp, qOverload<const shared_types::Regime&>(&JTTP::dataReceivedEvent), this, qOverload<const shared_types::Regime&>(&MainWindowManager::displayDataEvent));
     connect(jttp, qOverload<const shared_types::Control&>(&JTTP::dataReceivedEvent), this, qOverload<const shared_types::Control&>(&MainWindowManager::displayDataEvent));
@@ -28,35 +27,39 @@ void MainWindowManager::initialize() {
 
     // Request data for the first time on first time loading
     emit requestDataEvent(JTTP::REQUEST_DATA::SENSOR_DATA);
-    emit requestDataEvent(JTTP::REQUEST_DATA::REGIME_NAME);
     emit requestDataEvent(JTTP::REQUEST_DATA::REGIMES);
     emit requestDataEvent(JTTP::REQUEST_DATA::CONTROL_DATA);
+    emit requestDataEvent(JTTP::REQUEST_DATA::REGIME_ID);
 }
 
-void MainWindowManager::onMoistureSliderMoved(int value) {
-    control.getRegimeValue().setMoisture(value);
+void MainWindowManager::sendMoistureValue(int value) {
+    shared_types::RegimeValue regimeValue = control.getRegimeValue();
+    regimeValue.setMoisture(value);
+    control.setRegimeValue(regimeValue);
     sendData(control);
 }
 
-void MainWindowManager::onHeatSliderMoved(int value) {
-    control.getRegimeValue().setTemperature(value / MainWindow::SLIDER_MULTIPLIER);
+void MainWindowManager::sendHeatValue(int value) {
+    shared_types::RegimeValue regimeValue = control.getRegimeValue();
+    regimeValue.setTemperature(static_cast<float>(value) / MainWindow::SLIDER_MULTIPLIER);
+    control.setRegimeValue(regimeValue);
     sendData(control);
 }
 
-void MainWindowManager::onWindSliderMoved(int value) {
+void MainWindowManager::sendWindValue(int value) {
     control.setWindPercentage(value);
     sendData(control);
 }
 
-void MainWindowManager::onLightSliderMoved(int value) {
+void MainWindowManager::sendLightValue(int value) {
     control.setLightPercentage(value);
     sendData(control);
 }
 
-void MainWindowManager::onRegimeListChoice(const QString& currentText) {
-    shared_types::RegimeName regimeName;
-    regimeName.setName(currentText);
-    emit sendDataEvent(regimeName);
+void MainWindowManager::onRegimeListActivated(int index) {
+    shared_types::RegimeId regimeId;
+    regimeId.setId(index);
+    emit sendDataEvent(regimeId);
     emit requestDataEvent(JTTP::REQUEST_DATA::REGIME);
 }