+++ src/anticarium_desktop/MainWindowManager.h
@@ -1,5 +1,4 @@
-#ifndef MAINWINDOWMANAGER_H
-#define MAINWINDOWMANAGER_H
+#pragma once
 
 #include <anticarium_desktop/Jttp.h>
 #include <shared_types/Control.h>
@@ -13,27 +12,27 @@ class MainWindowManager : public QObject {
     MainWindowManager(QObject* parent = nullptr);
     void sendData(const shared_types::Control& control);
     void initialize();
+
+    void sendMoistureValue(int value);
+    void sendHeatValue(int value);
+    void sendWindValue(int value);
+    void sendLightValue(int value);
+
   signals:
     void sendDataEvent(const shared_types::Control& control);
-    void sendDataEvent(const shared_types::RegimeName& regimeName);
+    void sendDataEvent(const shared_types::RegimeId& regimeId);
     void requestDataEvent(JTTP::REQUEST_DATA requestData);
     void displayDataEvent(const shared_types::SensorData& newSensorData);
     void displayDataEvent(const shared_types::Regimes& newRegimes);
     void displayDataEvent(const shared_types::Control& newControl);
-    void displayDataEvent(const shared_types::RegimeName& newRegimeName);
+    void displayDataEvent(const shared_types::RegimeId& newRegimeId);
     void displayDataEvent(const shared_types::Regime& newRegime);
-  public slots:
-    void onMoistureSliderMoved(int value);
-    void onHeatSliderMoved(int value);
-    void onWindSliderMoved(int value);
-    void onLightSliderMoved(int value);
-    void onRegimeListChoice(const QString& currentText);
 
+  public slots:
+    void onRegimeListActivated(int index);
     // Stores incoming Control data
     void onDataReceived(const shared_types::Control& control);
 
   private:
     shared_types::Control control;
 };
-
-#endif // MAINWINDOWMANAGER_H