+++ src/anticarium_desktop/widgets/MainWindow.h
@@ -17,19 +17,31 @@ class MainWindow : public QMainWindow {
     ~MainWindow();
     static const int SLIDER_MULTIPLIER = 10;
   public slots:
-    void displayData(const shared_types::Regimes& regimes);
+    // Gets called only once when app starts
     void displayData(const shared_types::Control& control);
+
+    void displayData(const shared_types::Regimes& regimes);
     void displayData(const shared_types::SensorData& sensorData);
-    void displayData(const shared_types::RegimeName& regimeName);
+    void displayData(const shared_types::RegimeId& regimeId);
     void displayData(const shared_types::Regime& regime);
     void displayData(const shared_types::RegimeValue& regimeValue);
   private slots:
-    void onEnableSaveButton(int value);
+    void onEnableSaveButton(int value = 0);
+
+    // For slider value label
     void onMoistureSliderMoved(int value);
     void onHeatSliderMoved(int value);
     void onWindSliderMoved(int value);
     void onLightSliderMoved(int value);
-    void openRegimeDialog();
+
+    // For data sending
+    void onMoistureSliderReleased();
+    void onHeatSliderReleased();
+    void onWindSliderReleased();
+    void onLightSliderReleased();
+
+    void onOpenRegimeDialog();
+    void onOpenDisplayRegimes();
 
   private:
     Ui::MainWindow* ui;
@@ -43,7 +55,4 @@ class MainWindow : public QMainWindow {
 
     // Disonnects all ui input components that are used for data sending
     void disconnectUiInputs();
-
-    // Finds and sets element in regime list that corresponds to current regime label
-    void updateRegimeList();
 };